refactor(lib): extract run_diff orchestration into src/run.rs
src/lib.rs went from 1341 lines (47 KB) to 31 lines: it now only
declares the public module tree and re-exports the orchestration
entry points. All other code -- run_diff and its private helpers,
the public predicates (tripped, any_kev, any_epss_at_or_above,
budget_tripped), the public entry point run(), the embedded
INIT_* config templates, the lib tests -- moved verbatim into
the new src/run.rs.

Behavior-preserving: every consumer keeps its historical paths
(bomdrift::run, bomdrift::tripped, ...) via re-exports in lib.rs.
Test count unchanged (432 tests, all green); fmt + clippy 1.88
strict + cargo test --release --all-features all pass.
